Move Minesweeper specific files from api/util
