+++ src/main/java/ru/timeconqueror/lootgames/minigame/minesweeper/MSBoard.java
@@ -11,7 +11,6 @@ import java.util.stream.IntStream;
 import net.minecraft.nbt.NBTTagByte;
 import net.minecraft.nbt.NBTTagCompound;
 
-import ru.timeconqueror.lootgames.api.util.BombPerturbation;
 import ru.timeconqueror.lootgames.api.util.Pos2i;
 import ru.timeconqueror.lootgames.utils.future.ICodec;
 import ru.timeconqueror.timecore.api.util.CodecUtils;
