+++ src/main/java/ru/timeconqueror/lootgames/minigame/minesweeper/MSBoardSolver.java
@@ -10,9 +10,6 @@ import java.util.TreeMap;
 import java.util.TreeSet;
 
 import ru.timeconqueror.lootgames.LootGames;
-import ru.timeconqueror.lootgames.api.util.BombPerturbation;
-import ru.timeconqueror.lootgames.api.util.LogicMineSet;
-import ru.timeconqueror.lootgames.api.util.MineSets;
 import ru.timeconqueror.lootgames.api.util.Pos2i;
 
 /*
